% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofilter-package.R
\docType{package}
\name{geofilter-package}
\alias{geofilter-package}
\alias{geofilter}
\title{geofilter}
\description{
geofilter
}
\section{Package API}{

There are two main methods:
\itemize{
 \item \code{\link{sift_client}} - low level client
}
}

\examples{
# low level client, basically just using jqr
library(jsonlite)
x <- geofilter_data$filter_features
prettify(x)
sift_client(x, ".")
sift_client(x, "keys")

# high level client, using lazy eval
file <- system.file("examples", "zillow_or.geojson", package = "geofilter")
json <- paste0(readLines(file), collapse = "")
res <- sifter(json, COUNTY == Multnomah)
res \%>\% jqr::index() \%>\% jqr::dotstr(properties.COUNTY)
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
